feat(keyboard): First working version
First working version. There are a lot of things left to figure out. Integration with tinyusb is partly still makeshift and needs cleaning up. Tapdance is working, but it is possible that it still needs a minor change in architecture, so as to introduce an option to interrupt when pressing another button, and the hold handling needs attention