fix(keymaper): Fix FinalizeMappingRule and TapDanceMappingRule.
FinalizeMappingRule finalize only the pressed events. Other types must be finalized in specialized rules like TapDanceMappingRule.

TapDance now break a dance and publish keycode if other keycode is pressed. This function is good to typing. Maybe in future will be some settings is this keycode may be break, or not.